requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
uvicorn==0.30.6
Werkzeug==3.0.4
Flask-Session==0.8.0
python-dotenv